pyproject.toml
[tool.poetry]
name = "miky-bot"
version = "0.1.0"
description = ""
authors = ["mc_fdc <Masato.04.11.2007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.1"
aiohttp = "^3.8.5"
aiomysql = "^0.2.0"
jishaku = "^2.5.1"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
